Add the --nice argument to make
I call make quite regularly, and I almost always want to invoke it with
at least a bit of nice to avoid it eating my whole machine.  Since I can
stick "-j$(nproc)" in my shell's default MAKEFLAGS I frequently end up
being lazy and just typing "make", which I then have to kill in order to
re-run as "nice -n10 make".

That's a lot of typing, so this patch adds a "--nice" argument that,
when stuck in my makeflags, allows me to be lazy and avoid trashing my
whole machine.  I've never contributed to make before, so I'm not sure
what to do in terms of collateral -- I've just modified the usage text
and make.1.

* main.c (arg_nice): New variable.
  (usage): Add --nice argument.
  (switches): Add --nice argument.
  (main): Check arg_nice, and call nice() on POSIX systems.  Print a
  warning on non-POSIX systems or when nice() had no effect.